Reorg csv api
