+++ src/api/csvToXml/aggregationDisaggregationXml.ts
@@ -4,7 +4,7 @@ const createSingleAggregationDisaggregationEvent = (
   d: CsvAggregationDisaggregationPayloadFields
 ) => {
   return `                         
-<AggregationEvent> <!-- Aggregation  -->
+<AggregationEvent>
     <eventTime>${d.eventTime}</eventTime>
     <eventTimeZoneOffset>${d.eventTimeZoneOffset}</eventTimeZoneOffset>
     <parentID>${d.parentEPC}</parentID>
@@ -42,7 +42,7 @@ const createSingleAggregationDisaggregationEvent = (
 </AggregationEvent>`
 }
 
-export const createAggregationDisaggregationXml = async (
+export const createAggregatedAggregationDisaggregationXml = async (
   dList: [CsvAggregationDisaggregationPayloadFields]
 ) => {
   const d = dList[0]
@@ -52,8 +52,11 @@ export const createAggregationDisaggregationXml = async (
   <!-- GDST Basic Interoperability CSV File Nemo Converter - Aggregation / Disaggregation -->
   <epcis:EPCISDocument xmlns:epcis="urn:epcglobal:epcis:xsd:1" 
       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" 
-      xmlns:sbdh="http://www.unece.org/cefact/namespaces/StandardBusinessDocumentHeader" schemaVersion="0" creationDate="2001-12-17T09:30:47Z" xsi:schemaLocation="urn:epcglobal:epcis:xsd:1  http://www.gs1si.org/BMS/epcis/1_2/EPCglobal-epcis-1_2.xsd" 
-      xmlns:cbvmda="urn:epcglobal:cbv:mda" 
+      xmlns:sbdh="http://www.unece.org/cefact/namespaces/StandardBusinessDocumentHeader" 
+      schemaVersion="0" 
+      creationDate="2001-12-17T09:30:47Z" 
+      xsi:schemaLocation="urn:epcglobal:epcis:xsd:1  http://www.gs1si.org/BMS/epcis/1_2/EPCglobal-epcis-1_2.xsd" 
+      xmlns:cbvmda="urn:epcglobal:cbv:mda"
       xmlns:gdst="https://traceability-dialogue.org/epcis">
       <EPCISHeader>
           <sbdh:StandardBusinessDocumentHeader>
