+++ src/mocks/aggDisagg.csv
@@ -1,1 +1,2 @@
 informationProviderID,informationProviderContact,informationProviderEmail,informationReceiverID,informationReceiverContact,informationReceiverEmail,childEPC,informationProviderChild,speciesCodeChild,speciesNameChild,descriptionShortChild,itemConditionCodeChild,parentEPC,informationProviderParent,descriptonShortParent,speciesCodeParent,speciesNameParent,preservationTechniqueCode,unitOfMeasure,grossWeight,itemConditionCodeParent,additionalTradeItemIdentificationParent,location,name,partyIDTypeCode,additionalPartyID,streetAddressOne,streetAddressTwo,city,state,postalCode,countryCode,latitude,longitude,contact,telephone,email,action,eventTime,eventTimeZoneOffset,readPoint,bizLocation,purchaseOrderNumber,quantity,productOwner,informationProvider
+informationProviderID,informationProviderContact,informationProviderEmail,informationReceiverID,informationReceiverContact,informationReceiverEmail,childEPC,informationProviderChild,speciesCodeChild,speciesNameChild,descriptionShortChild,itemConditionCodeChild,parentEPC,informationProviderParent,descriptonShortParent,speciesCodeParent,speciesNameParent,preservationTechniqueCode,unitOfMeasure,grossWeight,itemConditionCodeParent,additionalTradeItemIdentificationParent,location,name,partyIDTypeCode,additionalPartyID,streetAddressOne,streetAddressTwo,city,state,postalCode,countryCode,latitude,longitude,contact,telephone,email,action,eventTime,eventTimeZoneOffset,readPoint,bizLocation,purchaseOrderNumber,quantity,productOwner,informationProvider
